+++ src/app/clientes/clientes.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ClientesService } from '../services/clientes.service';
 import { MatDialog } from '@angular/material';
 import { HelpComponent } from '../modals/help/help.component';
+import { TranslateService } from '../translate/translate.service';
 
 @Component({
   selector: 'app-clientes',
@@ -29,7 +30,8 @@ export class ClientesComponent implements OnInit {
   isLoading = false;
 
   constructor( protected clientesService: ClientesService,
-              public dialog: MatDialog ) { }
+              public dialog: MatDialog,
+              protected translateService: TranslateService ) { }
 
   ngOnInit() { this.readClientes(); }
 
@@ -62,6 +64,7 @@ export class ClientesComponent implements OnInit {
   }
 
   remove(id: number) {
+    if ( !confirm(this.translateService.instant('deleteWarning')) ) return ;
     this.clientesService.deleteCliente(id).then( (clientes) => {
       this.clientes = clientes;
       this.updateClientesToShow();
